+++ commerce03/src/main/java/com/supercoding/commerce03/web/advice/LoggingAspect.java
@@ -1,4 +1,4 @@
-package com.supercoding.commerce03.web.filter;
+package com.supercoding.commerce03.web.advice;
 
 import javax.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
